moved FilmInfo to it's own file, deleted the DataManager class
