+++ StarWars/StarWars/Model/FilmInfo.swift
@@ -1,0 +1,15 @@
+//
+//  FilmInfo.swift
+//  StarWars
+//
+//  Created by Lauren Small on 12/3/18.
+//  Copyright © 2018 Lauren Small. All rights reserved.
+//
+
+import Foundation
+
+
+
+struct FilmInfo: Codable {
+    var characters: [URL]
+}
